Remove conditions 1.0 component
Refactored some helper code into different locations so that all top condition files under "content" could be removed.
